+++ Glass/utils.lua
@@ -15,3 +15,9 @@ Utils.print = function (str, t)
     table.insert(Core.printBuffer, {str, t})
   end
 end
+
+---
+-- Prints Glass' notification messages
+Utils.notify = function (message)
+  print("|c00DFBA69Glass|r: ", message)
+end