+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Force chatStyle to classic (#94)
+
 ## 1.3.0-alpha2 (2020-09-06)
 
 * Improve scrolling behavior (#92)